refactor : review model-domain 분리
